+++ web/src/stylist/pages/VerifyCodePage.tsx
@@ -1,8 +1,0 @@
-import { useAppDispatch } from "../store";
-import { goToScreen } from "../store/consultationSlice";
-import { SignInPage } from "./SignInPage";
-
-export function VerifyCodePage() {
-  const dispatch = useAppDispatch();
-  return <SignInPage onBack={() => dispatch(goToScreen("sign-in"))} />;
-}